Move `KVEventsConfig` from `config/__init__.py` to `config/kv_events.py`
